Add reverse sync measurement method.
AVnu specifies a measurement method to remotely measure the clock
quality. Basically, this method sends Sync and/or Follow-up messages on
a slave towards the master, which can capture them and measure the
quality.

The AVnu specification [1] introduces three parameters:
 - reverseSyncEnabled
 - reverseSyncDomain
 - reverseSyncRate

The reverse sync messages can be sent in another domain to ensure they
don't conflict with the real sync messages. The rate is specified as a
fraction of the rate of the incoming sync messages. By default this is
95% of the incoming rate to make sure the sample interval is distributed
between two synchronization points. Additionally, the timer is free
running and not locked to the incoming sync interval.

The three paramenters are implemented as a per port property.

[1] https://avnu.org/wp-content/uploads/2014/05/Avnu-Testability-802.1AS-Recovered-Clock-Quality-Measurement-1.0_Approved-for-Public-Release.pdf

Signed-off-by: Michael Walle <michael@walle.cc>